feat: MyPage NetworkResponse DTO SSNetwork로 옮김
